// src/types/EditarUserType.ts
/* eslint-disable prettier/prettier */
export interface EditarUserType {
  descricao: string;
  fone: string;
  email: string;
  senha: string;
  cnpj: string;
  celular: string;
  tipoprof: string;
  percomprof: string;
  uf: string;
  dtnasc: string;
  rg_ie: string;
  profissao: string;
  bairro: string;
  cep: string;
  cidade: string;
  endereco: string;
  dtcadastro: string;
}